+++ README.md
@@ -1,9 +1,9 @@
-# sony-ci-api
+# Sony Ci API
 
-[![Gem Version](https://badge.fury.io/rb/sony-ci-api.svg)](http://badge.fury.io/rb/sony-ci-api)
-[![Build Status](https://travis-ci.org/WGBH/sony-ci-api.svg)](https://travis-ci.org/WGBH/sony-ci-api)
+[![Gem Version](https://badge.fury.io/rb/sony_ci_api.svg)](http://badge.fury.io/rb/sony_ci_api)
+[![Build Status](https://travis-ci.org/WGBH/sony_ci_api.svg)](https://travis-ci.org/WGBH/sony_ci_api)
 
-Provides a [Ruby interface](http://www.rubydoc.info/gems/sony-ci-api) 
+Provides a [Ruby interface](http://www.rubydoc.info/gems/sony_ci_api) 
 to the [Sony Ci REST API](http://developers.cimediacloud.com/). Features:
 - Takes care of OAuth.
 - Single method handles upload of any size file.
@@ -18,7 +18,7 @@ based on the sample in the same directory.
 ## Command-line usage
 
 ```
-$ sony-ci-api # lists the available options
+$ sony_ci_api # lists the available options
 ```
 
 This should work in any project which includes the gem, when run from the root of
@@ -27,7 +27,7 @@ the project, with `config/ci.yml` in place on that project.
 ## Library usage
 
 ```
-$ irb -Ilib -rsony-ci-api
+$ irb -Ilib -rsony_ci_api
 > ci = SonyCiAdmin.new(credentials_path: 'config/ci.yml')
 > ci.list_names # etc.
 ```
@@ -46,5 +46,5 @@ To publish gem:
 curl -u my-user-name https://rubygems.org/api/v1/api_key.yaml > ~/.gem/credentials
 chmod 0600 ~/.gem/credentials
 ```
-- Create gem: `gem build sony-ci-api.gemspec`
-- Push gem: `gem push sony-ci-api-X.Y.Z.gem`
+- Create gem: `gem build sony_ci_api.gemspec`
+- Push gem: `gem push sony_ci_api-X.Y.Z.gem`
