+++ bin/sony_ci_api
@@ -1,6 +1,6 @@
 #!/usr/bin/env ruby
 
-require_relative '../lib/sony-ci-api/sony_ci_admin'
+require_relative '../lib/sony_ci_api/sony_ci_admin'
 
 args = begin
   Hash[ARGV.slice_before { |a| a.match(/^--/) }.to_a.map { |a| [a[0].gsub(/^--/, ''), a[1..-1]] }]
