--- lib/sony-ci-api.rb
@@ -1,4 +1,0 @@
-require_relative 'sony-ci-api/sony_ci_admin'
-# Nothing else: The outside world doesn't need to worry
-# about the particular classes which are required, and
-# sony_ci_admin.rb loads its own dependencies.
