+++ lib/sony_ci_api/sony_ci_admin.rb
