+++ lib/sony_ci_api/sony_ci_basic.rb
