+++ lib/sony_ci_api/sony_ci_client.rb
