+++ sony_ci_api.gemspec
@@ -1,13 +1,13 @@
 Gem::Specification.new do |s|
-  s.name        = 'sony-ci-api'
+  s.name        = 'sony_ci_api'
   s.version     = '0.2.0'
-  s.executables << 'sony-ci-api'
+  s.executables << 'sony_ci_api'
   s.date        = '2015-10-06'
   s.summary     = 'Sony Ci API'
   s.description = 'Wrapper for the Sony Ci API (http://developers.cimediacloud.com/)'
-  s.authors     = ['Chuck McCallum']
-  s.email       = 'chuck_mccallum@wgbh.org'
+  s.authors     = ['Chuck McCallum', 'Andrew Myers']
+  s.email       = 'andrew_myers@wgbh.org'
   s.files       = Dir.glob('lib/**/*')
-  s.homepage    = 'https://github.com/WGBH/sony-ci-api'
+  s.homepage    = 'https://github.com/WGBH/sony_ci_api'
   s.license     = 'MIT'
 end
